Fix CodeQL LLM import cycles
